pyproject.toml
[project]
name = "paper-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.53.0",
    "loguru>=0.7.3",
    "mcp>=1.9.3",
    "openai>=1.85.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "tiktoken>=0.9.0",
]
